refactor bar component
