+++ components/Bar.js
@@ -1,57 +1,14 @@
 import React from 'react';
 import { View, Text, StyleSheet } from 'react-native';
-
+import { genPlateComponents } from '../utilities/functions/functions';
 import Plate from './Plate';
 
 const Bar = ({ weight, thirtyFive }) => {
-  const plates = [];
+  //Create array of objects. Each object defines a plate Compnent
+  const plates = genPlateComponents(weight, thirtyFive);
 
-  const plateNumbers = {};
-
-  weight = weight - 45;
-
-  while (weight >= 90) {
-    //45lb Plates
-    plates.push({ color: '#0F52BA', width: 15, height: 65 });
-    plateNumbers['fortyFive'] = plateNumbers['fortyFive'] + 1 || 1;
-    weight = weight - 90;
-  }
-
-  if (thirtyFive) {
-    while (weight >= 70) {
-      plates.push({ color: '#FCF55F', width: 15, height: 65 });
-      plateNumbers['thirtyFive'] = plateNumbers['thirtyFive'] + 1 || 1;
-      weight = weight - 70;
-    }
-  }
-
-  while (weight >= 50) {
-    //25lb Plates
-    plates.push({ color: '#009E60', width: 15, height: 65 });
-    plateNumbers['twentyFive'] = plateNumbers['twentyFive'] + 1 || 1;
-    weight = weight - 50;
-  }
-
-  while (weight >= 20) {
-    //10lb plates
-    plates.push({ color: '#FAF9F6', width: 9, height: 43 });
-    plateNumbers['ten'] = plateNumbers['ten'] + 1 || 1;
-    weight = weight - 20;
-  }
-
-  while (weight >= 10) {
-    //5lb plates
-    plates.push({ color: '#880808', width: 9, height: 33 });
-    plateNumbers['five'] = plateNumbers['five'] + 1 || 1;
-    weight = weight - 10;
-  }
-
-  while (weight >= 5) {
-    //2.5lb plates
-    plates.push({ color: 'black', width: 6, height: 25 });
-    plateNumbers['twoPointFive'] = plateNumbers['twoPointFive'] + 1 || 1;
-    weight = weight - 5;
-  }
+  //Last object in plates array is an object of plates as keys and the number of each plate as values
+  const plateNumbers = plates[plates.length - 1];
 
   return (
     <View style={styles.barContainer}>
@@ -86,7 +43,7 @@ const Bar = ({ weight, thirtyFive }) => {
             {plateNumbers.fortyFive + ': 45lbs'}
           </Text>
         ) : null}
-        {/* //Render 35lb plate */}
+        {/* Render 35lb plate */}
         {plateNumbers.thirtyFive ? (
           <Text style={styles.plateNumbersText}>
             {plateNumbers.thirtyFive + ': 35lbs'}
@@ -156,8 +113,6 @@ const styles = StyleSheet.create({
     padding: 20,
     top: 100,
     position: 'absolute',
-    // borderColor: 'red',
-    // borderWidth: 1,
   },
 
   plateNumbersText: {
